Refactor. Collapse 'simple' variants
Reduces code duplication and removes the need for genericised Program.
Scanner, Disassembler, and Compiler are now much longer but less tangled.
