+++ ulox/ulox.core/Package/Runtime/Compiler/Compilettes/DependencyInjectionCompilette.cs
@@ -2,14 +2,14 @@
 {
     public class DependencyInjectionCompilette
     {
-        public void Inject(CompilerBase compiler, bool canAssign)
+        public void Inject(Compiler compiler, bool canAssign)
         {
             compiler.TokenIterator.Consume(TokenType.IDENTIFIER, "Expect property name after 'inject'.");
             byte name = compiler.AddStringConstant();
             compiler.EmitOpAndBytes(OpCode.INJECT, name);
         }
 
-        public void RegisterStatement(CompilerBase compiler)
+        public void RegisterStatement(Compiler compiler)
         {
             compiler.TokenIterator.Consume(TokenType.IDENTIFIER, "Must provide name after a register statement.");
             var stringConst = compiler.AddStringConstant();
