+++ ulox/ulox.core/Package/Runtime/Interfaces/IParseRule.cs
@@ -3,7 +3,7 @@
     public interface IParseRule
     {
         Precedence Precedence { get; }
-        void Prefix(CompilerBase compiler, bool canAssign);
-        void Infix(CompilerBase compiler, bool canAssign);
+        void Prefix(Compiler compiler, bool canAssign);
+        void Infix(Compiler compiler, bool canAssign);
     }
 }
