+++ ulox/ulox.core/Package/Runtime/Scanner/Scanner.cs
@@ -18,7 +18,78 @@ namespace ULox
 
         public Scanner()
         {
+            Setup();
             Reset();
+        }
+
+        private void Setup()
+        {
+            var identScannerGen = new IdentifierScannerTokenGenerator();
+
+            this.AddGenerators(
+                new WhiteSpaceScannerTokenGenerator(),
+                new StringScannerTokenGenerator(),
+                new NumberScannerTokenGenerator(),
+                new SlashScannerTokenGenerator(),
+                new CompoundCharScannerCharMatchTokenGenerator(),
+                identScannerGen
+                                    );
+
+            identScannerGen.Add(
+                ("var", TokenType.VAR),
+                ("string", TokenType.STRING),
+                ("int", TokenType.INT),
+                ("float", TokenType.FLOAT),
+                ("and", TokenType.AND),
+                ("or", TokenType.OR),
+                ("if", TokenType.IF),
+                ("else", TokenType.ELSE),
+                ("while", TokenType.WHILE),
+                ("for", TokenType.FOR),
+                ("loop", TokenType.LOOP),
+                ("return", TokenType.RETURN),
+                ("break", TokenType.BREAK),
+                ("continue", TokenType.CONTINUE),
+                ("true", TokenType.TRUE),
+                ("false", TokenType.FALSE),
+                ("null", TokenType.NULL),
+                ("fun", TokenType.FUNCTION),
+                ("throw", TokenType.THROW),
+                ("yield", TokenType.YIELD),
+                ("fname", TokenType.CONTEXT_NAME_FUNC),
+
+                ("test", TokenType.TEST),
+                ("testcase", TokenType.TESTCASE),
+
+                ("build", TokenType.BUILD),
+
+                ("tcname", TokenType.CONTEXT_NAME_TESTCASE),
+                ("tsname", TokenType.CONTEXT_NAME_TESTSET),
+
+                ("class", TokenType.CLASS),
+                ("mixin", TokenType.MIXIN),
+                ("this", TokenType.THIS),
+                ("super", TokenType.SUPER),
+                ("static", TokenType.STATIC),
+                ("init", TokenType.INIT),
+                ("cname", TokenType.CONTEXT_NAME_CLASS),
+                ("freeze", TokenType.FREEZE),
+
+                ("inject", TokenType.INJECT),
+                ("register", TokenType.REGISTER)
+                                              );
+
+            this.AddSingleCharTokenGenerators(
+                ('(', TokenType.OPEN_PAREN),
+                (')', TokenType.CLOSE_PAREN),
+                ('{', TokenType.OPEN_BRACE),
+                ('}', TokenType.CLOSE_BRACE),
+                (',', TokenType.COMMA),
+                (';', TokenType.END_STATEMENT),
+                ('.', TokenType.DOT),
+                (':', TokenType.COLON),
+                ('?', TokenType.QUESTION)
+                                                    );
         }
 
         public void AddGenerator(IScannerTokenGenerator gen)
@@ -91,5 +162,21 @@ namespace ULox
 
         public void AddToken(TokenType simpleToken, string str, object literal)
             => Tokens.Add(new Token(simpleToken, str, literal, Line, CharacterNumber));
+
+
+        private void AddSingleCharTokenGenerator(char ch, TokenType tt)
+            => AddGenerator(new ConfiguredSingleCharScannerCharMatchTokenGenerator(ch, tt));
+
+        private void AddSingleCharTokenGenerators(params (char ch, TokenType token)[] tokens)
+        {
+            foreach (var item in tokens)
+                AddSingleCharTokenGenerator(item.ch, item.token);
+        }
+
+        private void AddGenerators(params IScannerTokenGenerator[] scannerTokenGenerators)
+        {
+            foreach (var item in scannerTokenGenerators)
+                AddGenerator(item);
+        }
     }
 }
